Separate classes and parse input
